bluejay-battery: add battery replacement tip from stock OS
Via stock OS logic, the tip will only appear when the system reports that the battery health is
either

(1) fair and cycle count >= 375 (this will appear as an early warning)
(2) "dead"

Requires adevtool changes to add synthetic overlays for SettingsGoogle strings.

Requires an upstream kernel change for battery status to be updated accordingly on affected devices
(change is currently present in 16QPR1 kernels).